Created config validator Also moved config classes to config folder
